fix: 일러스트 랜덤 적용 및 최적화 적용
